+++ src/extension.ts
@@ -17,6 +17,7 @@ import { DependencyExplorer } from "./views/dependencyExplorer";
 export async function activate(context: ExtensionContext): Promise<void> {
     contextManager.initialize(context);
     await initializeFromJsonFile(context.asAbsolutePath("./package.json"), { firstParty: true });
+    await initExpService(context);
     await instrumentOperation("activation", activateExtension)(context);
     contextManager.setContextValue(Context.EXTENSION_ACTIVATED, true);
     contextManager.setContextValue(Context.SUPPORTED_BUILD_FILES, Build.FILE_NAMES);
@@ -30,7 +31,6 @@ async function activateExtension(_operationId: string, context: ExtensionContext
     context.subscriptions.push(contextManager);
     context.subscriptions.push(syncHandler);
     context.subscriptions.push(tasks.registerTaskProvider(ExportJarTaskProvider.exportJarType, new ExportJarTaskProvider()));
-    initExpService(context);
 }
 
 // this method is called when your extension is deactivated